final commit of virus spreading model

+++ #3-simple-model-for-virus-spreading/src/utils/init_config.c
@@ -1,6 +1,6 @@
 #include "init_config.h"
 
-init_config_t set_init_config(int total_people, int infected_people, int W, int L, int w, int l, int v, int d, int t){
+init_config_t set_init_config(int total_people, int infected_people, int W, int L, int w, int l, float v, int d, int t){
     init_config_t init_config;
     init_config.total_people = total_people;
     init_config.infected_people = infected_people;
